feat: move database function to another module
