+++ src/main.rs
@@ -1,20 +1,21 @@
 mod arg_parser;
+mod database;
 mod utils;
 use std::{
     io::{self, Write},
     process::exit,
 };
 
-use utils::Database;
-#[macro_use]
-extern crate prettytable;
+use database::Database;
 
 fn execute_delete_by_id(database: &Database) {
     println!();
     if utils::output_data(
-        &utils::get_all_entries(database).expect("Failed while trying to read all entries."),
+        &database
+            .get_all_entries()
+            .expect("Failed while trying to read all entries."),
         false,
-    ) <= 0
+    ) == 0
     {
         println!("No data to delete");
         exit(0);
@@ -24,23 +25,24 @@ fn execute_delete_by_id(database: &Database) {
     print!("Enter the id of the entry you want to remove: ");
     io::stdout().flush().unwrap();
     io::stdin().read_line(&mut id).expect("Failed to read line");
-    utils::delete_entry(
-        id.trim().parse::<i64>().expect("Failed to parse id"),
-        database,
-    );
+    database.delete_entry(id.trim().parse::<i64>().expect("Failed to parse id"));
 }
 
 fn execute_list(del_links: bool, database: &Database) {
     println!();
     utils::output_data(
-        &utils::get_all_entries(database).expect("Failed while trying to read all entries."),
+        &database
+            .get_all_entries()
+            .expect("Failed while trying to read all entries."),
         del_links,
     );
     println!();
 }
 
 fn execute_drop(database: &Database) {
-    utils::delete_database_file(database).expect("Failed to delete database file.");
+    database
+        .delete_database_file()
+        .expect("Failed to delete database file.");
 }
 
 fn execute_transfer(path: &str, database: &Database) {
@@ -68,10 +70,12 @@ fn execute_transfer(path: &str, database: &Database) {
             entry_name = default_name.to_string();
         }
         println!("\nUploading... please wait\n");
-        utils::transfer_file(entry_name.trim(), path, database);
+        database.transfer_file(entry_name.trim(), path);
     }
     utils::output_data(
-        &utils::get_all_entries(database).expect("Failed while trying to read all entries."),
+        &database
+            .get_all_entries()
+            .expect("Failed while trying to read all entries."),
         false,
     );
     println!();
@@ -80,7 +84,7 @@ fn execute_transfer(path: &str, database: &Database) {
 fn run_app() {
     let database = Database::new();
     utils::create_config_app_folder().expect("Failed to create config folder.");
-    utils::create_table(&database).expect("Failed to create table.");
+    database.create_table().expect("Failed to create table.");
     let args = arg_parser::prepare_args();
 
     match args {
